Reshuffled functions as appropriate between models and util
